add randomize for tiles rotation; add different initial game states.
